:truck: rename: _atom folder name into atom
